core/requirements-dev.txt
-r requirements.txt

black==18.9b0
coverage==4.5.2
ipdb==0.11
pylint==2.2.2
pylint-django==2.0.5